chore: remove redundant types and reorganize files
