--- packages/rolldown/src/constants/types.ts
@@ -1,1 +1,0 @@
-export type StringOrRegExp = string | RegExp
