+++ packages/rolldown/src/rolldown-build.ts
@@ -4,7 +4,7 @@ import { BundlerWithStopWorker, createBundler } from './utils/create-bundler'
 import type { InputOptions } from './types/input-options'
 import type { OutputOptions } from './types/output-options'
 import type { RolldownOutput } from './types/rolldown-output'
-import type { HasProperty, TypeAssert } from './utils/type-assert'
+import type { HasProperty, TypeAssert } from './types/assert'
 
 export class RolldownBuild {
   #inputOptions: InputOptions
