+++ packages/rolldown/src/utils/normalize-string-or-regex.ts
